Refactor appVersion duck
