+++ ducks/appVersion/common.ts
@@ -1,0 +1,2 @@
+import { actionTypeCreator } from '@/utils';
+export const $ = actionTypeCreator('APP_VERSION');
